openai-sdk-agents-assigments/07-Class_Input-guardial/pyproject.toml
[project]
name = "07-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.8",
    "rich>=14.1.0",
]
